requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.0.0
httpx>=0.25.0
pydantic>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
rank-bm25>=0.2.2
sentence-transformers>=2.2.2
tiktoken>=0.5.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
rich>=13.0.0
tqdm>=4.66.0
